requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
asynctest==0.13.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
iniherit==0.3.9
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tabulate==0.8.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
